Now the hotbar is ready to be use (Can be refacto easily if needed)
I change the use of the hotbar, i first wanted it to be separated from the main inventory, like a little inventory c=that can hold items but the problem is that it can created to much slot used for the same item for nothing. So what i made is the hotbar is now related to the main inventory but it can be changed anytime.